refactor: Move TextEncoding into its own file
