+++ python/arrow_odbc/read_arrow_batches_from_odbc.py
@@ -8,8 +8,8 @@ from .reader import (
     DEFAULT_FETCH_BUFFER_LIMIT_IN_BYTES,
     DEFAULT_FETCH_BUFFER_LIMIT_IN_ROWS,
     BatchReader,
-    TextEncoding,
 )
+from .text_encoding import TextEncoding
 
 
 def read_arrow_batches_from_odbc(
